+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Fix ``chmod`` command syntax and run it only if needed in ``Cache::clearTemporaryDirectory`` [PR-24](https://github.com/OXID-eSales/testing_library/pull/25)
+- Make module services available in the DI container
 
 ## [v7.1.2] - 2020-07-16